+++ encoding/json/outbound_test.go
@@ -51,7 +51,7 @@ func TestCall(t *testing.T) {
 
 	tests := []struct {
 		procedure       string
-		headers         yarpc.Headers
+		headers         map[string]string
 		body            interface{}
 		encodedRequest  string
 		encodedResponse string
@@ -61,7 +61,7 @@ func TestCall(t *testing.T) {
 
 		// Either want, or wantType and wantErr must be set.
 		want        interface{} // expected response body
-		wantHeaders yarpc.Headers
+		wantHeaders map[string]string
 		wantType    reflect.Type // type of response body
 		wantErr     string       // error message
 	}{
@@ -89,12 +89,12 @@ func TestCall(t *testing.T) {
 		},
 		{
 			procedure:       "requestHeaders",
-			headers:         yarpc.NewHeaders().With("user-id", "42"),
+			headers:         map[string]string{"user-id": "42"},
 			body:            map[string]interface{}{},
 			encodedRequest:  "{}",
 			encodedResponse: "{}",
 			want:            map[string]interface{}{},
-			wantHeaders:     yarpc.NewHeaders().With("success", "true"),
+			wantHeaders:     map[string]string{"success": "true"},
 		},
 	}
 
@@ -113,14 +113,14 @@ func TestCall(t *testing.T) {
 						Service:   service,
 						Procedure: tt.procedure,
 						Encoding:  Encoding,
-						Headers:   transport.Headers(tt.headers),
+						Headers:   transport.HeadersFromMap(tt.headers),
 						Body:      bytes.NewReader([]byte(tt.encodedRequest)),
 					}),
 			).Return(
 				&transport.Response{
 					Body: ioutil.NopCloser(
 						bytes.NewReader([]byte(tt.encodedResponse))),
-					Headers: transport.Headers(tt.wantHeaders),
+					Headers: transport.HeadersFromMap(tt.wantHeaders),
 				}, nil)
 		}
 
@@ -135,13 +135,11 @@ func TestCall(t *testing.T) {
 
 		var (
 			opts       []yarpc.CallOption
-			resHeaders yarpc.Headers
+			resHeaders map[string]string
 		)
 
-		for _, k := range tt.headers.Keys() {
-			if v, ok := tt.headers.Get(k); ok {
-				opts = append(opts, yarpc.WithHeader(k, v))
-			}
+		for k, v := range tt.headers {
+			opts = append(opts, yarpc.WithHeader(k, v))
 		}
 		opts = append(opts, yarpc.ResponseHeaders(&resHeaders))
 
@@ -176,7 +174,7 @@ func TestCallOneway(t *testing.T) {
 
 	tests := []struct {
 		procedure      string
-		headers        yarpc.Headers
+		headers        map[string]string
 		body           interface{}
 		encodedRequest string
 
@@ -198,7 +196,7 @@ func TestCallOneway(t *testing.T) {
 		},
 		{
 			procedure:      "requestHeaders",
-			headers:        yarpc.NewHeaders().With("user-id", "42"),
+			headers:        map[string]string{"user-id": "42"},
 			body:           map[string]interface{}{},
 			encodedRequest: "{}\n",
 		},
@@ -218,7 +216,7 @@ func TestCallOneway(t *testing.T) {
 					Service:   service,
 					Procedure: tt.procedure,
 					Encoding:  Encoding,
-					Headers:   transport.Headers(tt.headers),
+					Headers:   transport.HeadersFromMap(tt.headers),
 					Body:      bytes.NewReader([]byte(tt.encodedRequest)),
 				})
 
@@ -237,10 +235,8 @@ func TestCallOneway(t *testing.T) {
 
 		var opts []yarpc.CallOption
 
-		for _, k := range tt.headers.Keys() {
-			if v, ok := tt.headers.Get(k); ok {
-				opts = append(opts, yarpc.WithHeader(k, v))
-			}
+		for k, v := range tt.headers {
+			opts = append(opts, yarpc.WithHeader(k, v))
 		}
 
 		ack, err := client.CallOneway(ctx, tt.procedure, tt.body, opts...)